authorization

+++ modules/post/post.controller.js
@@ -1,38 +1,53 @@
-const Post = require ("./post.model")
+const Post = require ("./post.model");
+
+const verifyAuthor = async (req,user)=>{
+let post = await Post.findById(req.params.postId);
+if (post._id.toString()!== user.req.id){
+    return res.status(406).json({error:"You are not permitted to perform this operation"});
+}
+}
 
 exports.getPosts = async (req,res)=>{
     const posts = await Post.find({});
     res.status(200).json({posts});
+};
+
+exports.getPostsByUser = async (req,res)=>{
+    const posts = await Post.find({author: req.user.id});
+    res.status(200).json({posts});
 }
 
-exports.getPost = (req,res)=>{
-    res.status(200).send("Gets a single Post")
-}
 exports.createPost =async (req,res)=>{
+    const {title, body, published}= req.body
     const post = await Post.create({
-        title: req.body.title,
-        body: req.body.body,
-        published: req.body.published,
+        title,body,published,
+        author: req.user.id,
     });
-    res.status(201).send({post})
+    res.status(201).json({post})
 };
 
 exports.getPost = async (req,res)=>{
-    // const postId = req.params.postId;
     const {postId} = req.params;
     const post = await Post.findById(postId);
     res.status(200).json({post});
 }
 
-
 exports.updatePost=async (req,res)=>{
     const {postId} = req.params;
+
+    //checks
+    await verifyAuthor();
+
     const post = await Post.findByIdAndUpdate(postId,{...req.body},{new:true});
     res.status(200).json({post});
 }
 
 exports.deletePost= async(req,res)=>{
     const {postId} = req.params;
+
+    //checks
+    await verifyAuthor();
+
     const post = await Post.findByIdAndDelete(postId,);
     res.status(200).json({msg:"Post Succesfully deleted."});
 }
